Separate api to the different files
